update v2
Modified the OS versions it was tested on. Added advisory for running Powershell in administrator mode. Added command to install current LTS version of Ubuntu. Added command to verify installed Ubuntu distribution. Added advisory to ensure that installation of Node and other packages is done in Ubuntu. Added command to install Node version between 16.0 and 18.0. Added instructions on certificates and address error with wss.